Review Hibernate 3.6.5 patch and OpenMRS upgrade patch - SEARCH-10 Author: czakian
git-svn-id: http://svn.openmrs.org/openmrs/trunk@24899 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/db/LocationDAO.java
@@ -145,7 +145,7 @@ public interface LocationDAO {
 	/**
 	 * @see org.openmrs.api.LocationService#getCountOfLocations(String, Boolean)
 	 */
-	public Integer getCountOfLocations(String nameFragment, Boolean includeRetired);
+	public Long getCountOfLocations(String nameFragment, Boolean includeRetired);
 	
 	/**
 	 * @see LocationService#getRootLocations(boolean)
